Add basic functionality to main
Added while loop that takes in input until bye is entered.

Also renamed Duke class to Internity.
