+++ src/test/java/seedu/duke/InternityTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class InternityTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
